add new timeline item. split up feature
